Day 1 Solution
Pretty simplistic day 1. Collecting all the lines into a Vector allows you to use the windows function to create necessary pairs for part 1 and sliding windows for part 2.  Only slightly annoying part was using fold in part 2 to hold our count and the previous sum to calculate whether to increment the sum or not. Its functional, just not the nicest looking solution.