Day 2 Solution
Simple solution using a tuple to hold each move. Part 2 kinda messes up the meaning of the tuple, but that the point of the prompt.

The lack of iterators really make this less appealing, but using reduce in part 1 was generating a lifetime issue. In the interest of time refactor into the more simplistic for each solution.